requirements.txt
Django==1.5.1
Markdown==2.3.1
South==0.8.1
distribute==0.6.31
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-tastypie==0.10.0
django-toolbelt==0.0.1
djangorestframework==2.3.6
gunicorn==17.5
psycopg2==2.5.1
python-dateutil==2.1
python-mimeparse==0.1.4
six==1.3.0
static==0.4
wsgiref==0.1.2
